backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
python-dotenv==1.0.0
python-multipart==0.0.6
openai==1.10.0
anthropic==0.18.0
pandas==2.2.0
openpyxl==3.1.2
aiohttp==3.9.3
